Simplify scheduling a Function and more refactors
