+++ packages/punchcard/lib/dynamodb/index.ts
@@ -1,3 +1,4 @@
+export * from './client';
 export * from './expression';
 export * from './key';
 export * from './query';
