+++ packages/punchcard/lib/dynamodb/query.ts
@@ -1,7 +1,8 @@
 import { BinaryType, RuntimeType, Shape, StringType, Type } from '../shape';
+import { Query } from './client';
 import { CompileContext } from './expression/compile-context';
 import { CompileContextImpl } from './expression/compiler';
-import { Query, Table } from './table';
+import { Table } from './table';
 
 export interface CompiledQuery {
   KeyConditionExpression: string;
