+++ packages/punchcard/lib/lambda/executor.ts
@@ -1,14 +1,14 @@
-import events = require('@aws-cdk/aws-events');
-import eventTargets = require('@aws-cdk/aws-events-targets');
 import lambda = require('@aws-cdk/aws-lambda');
 import cdk = require('@aws-cdk/core');
 
 import { Integration, LambdaIntegration, Resource } from '../api-gateway';
 import { Client } from '../core/client';
 import { Dependency } from '../core/dependency';
-import { Type } from '../shape/types/type';
 import { Omit } from '../util/omit';
-import { Function } from './function';
+import { Function, FunctionProps } from './function';
+
+import * as CloudWatch from '../cloudwatch';
+import { schedule, ScheduleProps } from './schedule';
 
 /**
  * Alias for creating a LambdaExecutorService
@@ -24,35 +24,18 @@ export class ExecutorService {
     memorySize: 128
   }) {}
 
-  public spawn<T, U, D extends Dependency<any>>(scope: cdk.Construct, id: string, props: {
-    request?: Type<T>,
-    response?: Type<U>,
-    depends: D,
-    handle: (event: T, run: Client<D>, context: any) => Promise<U>;
-  }): Function<T, U, D> {
+  public spawn<T, U, D extends Dependency<any>>(scope: cdk.Construct, id: string, props: FunctionProps<T, U, D>): Function<T, U, D> {
     return new Function<T, U, D>(scope, id, {
       ...this.props,
       ...props
     });
   }
 
-  public schedule<D extends Dependency<any>>(scope: cdk.Construct, id: string, props: {
-    schedule: events.Schedule;
-    depends?: D;
-    handle: (event: CloudwatchEvent, run: Client<D>, context: any) => Promise<any>;
-  }): Function<CloudwatchEvent, any, D> {
-    scope = new cdk.Construct(scope, id);
-    const f = new Function<CloudwatchEvent, any, D>(scope, 'Function', {
+  public schedule<D extends Dependency<any>>(scope: cdk.Construct, id: string, props: ScheduleProps<D>): Function<CloudWatch.Event, any, D> {
+    return schedule(scope, id, {
       ...this.props,
       ...props
     });
-
-    new events.Rule(scope, 'Schedule', {
-      schedule: props.schedule,
-      targets: [new eventTargets.LambdaFunction(f)]
-    });
-
-    return f;
   }
 
   public apiIntegration<D extends Dependency<any>>(parent: cdk.Construct, id: string, props: {
@@ -72,15 +55,4 @@ export class ExecutorService {
     const integration = new LambdaIntegration(handler);
     return integration;
   }
-}
-
-export interface CloudwatchEvent {
-  account: string;
-  region: string;
-  detail: any;
-  'detail-type': string;
-  source: string;
-  time: string;
-  id: string;
-  resources: string[];
 }
