+++ packages/punchcard/lib/sqs/index.ts
@@ -1,4 +1,4 @@
 export * from './collector';
 export * from './event';
-export * from './stream';
+export * from './messages';
 export * from './queue';
